scripts/grades/requirements.txt
openpyxl
pyxlsb
python-dotenv